requirements.txt
# Web Scraping
curl-cffi>=0.5.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Data & ML
pandas>=2.0.0
numpy>=1.24.0,<2.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# MLOps
mlflow>=2.9.0
dvc[s3]>=3.30.0
prefect>=2.14.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# NLP & Embeddings
torch>=2.1.0
sentence-transformers>=3.0.0
transformers>=4.35.0
anthropic>=0.39.0

# Utils
python-dotenv>=1.0.0
loguru>=0.7.0
pyyaml>=6.0.1

# Testing & Quality
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.5.0

# Data Quality
great-expectations>=0.18.0